titiler-vercel/requirements-alt.txt
# Core dependencies for Titiler
titiler.core==0.11.7
titiler.application==0.11.7
fastapi==0.95.2
starlette==0.27.0
uvicorn==0.22.0
python-dotenv==1.0.0

# Use pre-built GDAL and rasterio wheels from PyPI
GDAL==3.4.3
rasterio==1.3.7

# Additional dependencies for COG processing
rio-tiler==4.1.10
pydantic<2.0.0
geojson-pydantic==0.5.1
jinja2==3.1.2
python-multipart==0.0.6
simplejson==3.19.1

# AWS S3/R2 support
boto3==1.28.25
botocore==1.31.25
